Datasetv2
Non-overlapping images patches of 96x96 pixels were extracted from each binary image. Images with white pixels (cracks) over 5% are classified as crack images and less than 5% are no-crack images. A total of 409,432 images of 256,115 crack images and 153,316 no-crack images were obtained from the 5 videos processed. This resulting dataset is addressed as CRACK_V2 in the rest of this paper and it will be used as ground truth to train and test the segmentation models